Add XY table, beginnings of refactor to make table subordinate to plots (#19)
Modifies the XY plot widget to be a splitter with a table component.
Architecturally, this has the table subordinate to plots, the table is
only a controller into the plots and uses the signals from the plot
(instead of the top-level splitter being the coordinator). A bit of a
test of how this works before refactoring the main plot / signals-table
to use this architecture.

Changes the behavior of the top-level window (for the CSV viewer) to
close sub-XY-windows.

Changes to the multi-plot widget:
- Adds sigDataItemsUpdated, sigDataUpdated, which the XY plot listens on
directly (instead of through the table's data-updated)
- Add tests for updated signal
- Changes signals to fire only after the internal state has been
updated, e.g., region-updated signal after plots._region set

Changes to signals table:
- Base signals table now directly has a reference to the plots
- _render_value plumbed to the plots directly

Changes to XY plot architecture:
- Create a BaseXyPlot that can be inherited by the plot only or the
splitter.
- Provides a sigXysChanged signals, which can be used in combination
with plots.sigDataItemsChanged for a table to listen for updates
- Breaks out the drag-and-drop target feature into a separate mixin
- Split into three files: the base plot + widget, the SignalsTable item
to open XY plots, and the XY splitter widget
- Listens directly to the plot for data / region changes