Add reference geometry for xy plots, refactor save-model infrastructure (#20)
Add user-specified (as a function -> (xs, ys)) reference geometry,
available through the XY table menu. Modeled off data transforms, using
simpleeval for the functions, and with the option to modify / delete
existing reference geometry.

Some refactoring of the XY plot widget and components:
- Add ref-geo as a mixin into XyPlotWidget, and table-display / controls
as mixin into XyPlotTable
- Adds a XY-items changed signal
- Split into 3 files: base XY plot widget + controls table, splitter,
and SignalsTable item that provides for XY widget creation
- Add context menu mixin to XyPlotTable

Also a major infrastructural change to support mixin-save-models for
XyPlots. Splits the HasSaveLoadConfig into HasSaveLoadConfig (that is
structure-agnostic - for the XyPlot window) and HasSaveLoadDataConfig
(that imposes the top-level 'data' item, continue to be used for
top-level MultiPlotWidget / SignalsTable).
- Also changes the class variable to internal `_MODEL_BASES`
- Instead of returning the data type, that is now obtained by inspection
into the top-level model
- Requires a change in signature in write_model / load_model to
BaseModel instead of BaseTopModel, which is probably cleaner anyways
since most uses need to assert on the model fragment type
- Separate out _dump_model (which takes no arguments) and
_dump_data_model (a HasSaveLoadDataConfig which populates the data
field)

It is likely the HasSaveLoadConfig infrastructure will continue to
evolve, as XyPlots end up doing even more.

Other changes:
- Add sigDataUpdated and sigDataItemsUpdated to MultiPlotWidget and
associated tests
-