Restructure signals table to be a control into the plots (#22)
Previously, the signals table needed the top-level splitter to
coordinate between it and the plots. This makes the signal table operate
directly on the plots, eliminating the need for top-level boilerplate.

This is a hefty refactor. A major part of this is splitting files.

For widgets where the main logic was in the table (like timeshift or
transforms), moves the main logic into the plots and exposes an API for
the UI logic in the table. Some other logic previously at the top-level
(like creating immutable arrays) is moved into the plots.

Other:
- Allows _PLOT_TYPE / _TABLE_TYPE to specify the plot type instead of
requiring the more verbose _make_plot
- Add stats tests
- Adds base infrastructure for transforms in MultiPlotWidget, as
something between data-in and data-rendered. This facility accommodates
function-transforms and timeshifts.
- Allow _make_controls => None in top-level