Add visibility toggles to timeseries and xy plots (#24)
This follows a similar architecture to the recent refactor: an API for
visibility + backing model is in the plot widget, and the controls is in
the table. One set of plot-widget mixin and table mixin for each of
timeseries and xy plots. Adds these to the CSV viewer.

The backing model stores hidden state as the set hidden names.

Infrastructure changes:
- Default item flags exclude ItemIsUserCheckable, that is now used to
detect uninitialized state
- Store data names -> curves for both timeseries and xy plots, this is
the base infrastructure that allows visibility without full re-creation
- Refactor out the mixins col infrastructure (which allows mixins to
contribute columns to the overall table, and they automagically compose)
to MixinColsTable
  - Add this infrastructure to XyPlotTable
- XyTable now properly creates its save-state data model type from its
plot widget type
- Save-restore-model: add _create_class_model_bases as a dynamic
alternative to _MODEL_BASES, which is passed in its top-level cls
- rename _get_model_bases to _get_all_model_bases, to differentiate from
the new dynamic model base function
- BaseXyPlot: change closed -> sigClosed to avoid a name collision