Live stats on updating data (#33)
Changes the behavior of the stats table to be live on updating data.
More specifically, does not apply a 100ms delay and does not clear the
table (to avoid flicker) when data is updated, though those still happen
when the region is moved.

Also fixes the region tolerance code, to be a fraction of the region
width instead of the absolute values. That would break badly when used
with timestamps.

Architecturally, changes the cross-thread stats request from a queue to
a shared variable guarded by a mutex.