Use QThreadPool for stats calculator threading (#34)
destroyed and closeEvent are pretty unreliable for cleaning up threads,
so this uses a single-max-thread threadpool for thread management, and
the scope of each thread is to calculate once and be done.

This continues to use the single shared global state, so threads are not
independent.

A future PR may allow true multithreading, but this is intentionally
limited to one thread to make stats calculation as lightweight as
possible.