Add live filter mixin on signals table (#38)
As an alternative to the search mixin, add a filter mixin that has the
same textbox popup but live filters the rows shown on the table. On new
data items defined, the filter mixin clears. Changes the CSV viewer to
use this by default.

The filter uses new matching logic that finds all substrings (separated
by whitespace in the filter box) in table text, even if noncontiguous.

Also changes table behavior to no longer _update() in __init__(), since
in practice it does nothing (no data is defined, unless the table was
late-constructed which typically only happens in tests) and it calls
_update() before subclass __init__() have finished.