Nudge cursor with arrow keys (#53)
Adds a NudgeablePlot interactivity mixin, which adds nudge functionality
to the live cursor and region (if present). Keyboard left and right
arrows moves the current cursor (or the region end - if region is
mouseover'd) to the next (or previous) sample in any visible data set in
the plot widget. Widget focus is required.

Opportunistic cleanup: cleanup imports