Bugfix for color picker on load-config (#55)
Previously, load config does not set the table colors correctly for
signals with a color override set. Likely because color state was not
updated by the time the color table was updated. This restructures the
color code to replace the data into show_data_items so everything
downstream has the right data.

Also adds unit tests.